Create Defining and Using Attributes in a Python Class
This Python code demonstrates the concept of attributes within a class using the example of a Dog class. Attributes, such as name and age, represent specific properties of instances created from the class. The code showcases the initialization of attributes, creating instances of the class, and accessing attribute values using the dot notation. Attributes play a crucial role in encapsulation, enabling the bundling of data and behavior within a class for more organized and maintainable code.